Solucionados errores de la it-2(enumerados, atributos derivados para rendimiento, maxPersonas null en la bd)
Tomcat funcionando, cambiando ws-app-ds del ModelConstants a ws-javaexamples-ds , lo mismo en el pom.xml del ws-app

Rest funcional a medio implementar

Eliminado fichero de Mock que heredaba del ClientOfertaService

git-svn-id: https://svn.fic.udc.es/grao3/isd/13-14/isg016@37 dd163d99-12d2-43d0-9818-c3a854d4ac74

+++ ws-app/trunk/ws-app-service/src/main/java/es/udc/ws/app/soapservice/SoapOfertaMaxPersonasExceptionInfo.java
@@ -4,13 +4,13 @@ package es.udc.ws.app.soapservice;
 public class SoapOfertaMaxPersonasExceptionInfo {
 
     private Long ofertaId;
-    private short maxPersonas;
+    private Long maxPersonas;
 
     public SoapOfertaMaxPersonasExceptionInfo() {
     }
 
     public SoapOfertaMaxPersonasExceptionInfo(Long ofertaId, 
-                                           short maxPersonas) {
+                                           Long maxPersonas) {
         this.ofertaId = ofertaId;
         this.maxPersonas = maxPersonas;
     }
@@ -19,11 +19,11 @@ public class SoapOfertaMaxPersonasExceptionInfo {
         return ofertaId;
     }
 
-    public short getMaxPersonas() {
+    public Long getMaxPersonas() {
         return maxPersonas;
     }
 
-    public void setMaxPersonas(short maxPersonas) {
+    public void setMaxPersonas(Long maxPersonas) {
         this.maxPersonas = maxPersonas;
     }
